Update ControllerScript
this one extra line replaces setting the velocity directly with a creating an impulse force required to get to the desired velocity. This can be more stable than setting velocity directly especially if your character is walking into and bumping into geometry or pushing around other dynamic bodies. It should behave indistinguishably from the way as you had it before as well.